Clean up components for sign in page graphics
